# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "description": "This action returns all cats",
        "operationId": "getCats",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "This action returns all cats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CatEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all cats",
        "tags": [
          "cats"
        ]
      },
      "post": {
        "operationId": "createCat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The record has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatEntity"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "cats"
        ]
      }
    },
    "/{id}": {
      "get": {
        "operationId": "getCatsById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "This action returns a cat by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatEntity"
                }
              }
            }
          },
          "404": {
            "description": "The record with the given id was not found."
          }
        },
        "summary": "",
        "tags": [
          "cats"
        ]
      },
      "put": {
        "operationId": "updateCat",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The record has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCatDto"
                }
              }
            }
          },
          "404": {
            "description": "The record with the given id was not found."
          }
        },
        "summary": "",
        "tags": [
          "cats"
        ]
      },
      "delete": {
        "operationId": "deleteCat",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The record has been successfully deleted."
          },
          "404": {
            "description": "The record with the given id was not found."
          }
        },
        "summary": "",
        "tags": [
          "cats"
        ]
      }
    }
  },
  "info": {
    "title": "Cats example",
    "description": "The cats api documentation",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "cats",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "CatEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The cat id",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "The cat name",
            "example": "Kitty"
          },
          "sex": {
            "type": "string",
            "description": "The cat sex",
            "example": "MALE"
          },
          "age": {
            "type": "number",
            "description": "The cat age",
            "example": 1
          },
          "breed": {
            "type": "string",
            "description": "The cat breed",
            "example": "Persian"
          },
          "color": {
            "type": "string",
            "description": "The cat color",
            "example": "White"
          },
          "ownerId": {
            "type": "number",
            "description": "The cat owner id",
            "example": 1
          }
        },
        "required": [
          "id",
          "name",
          "sex",
          "age",
          "breed",
          "color",
          "ownerId"
        ]
      },
      "CreateCatDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The cat name",
            "example": "Kitty"
          },
          "sex": {
            "type": "string",
            "description": "The cat sex",
            "example": "MALE"
          },
          "age": {
            "type": "number",
            "description": "The cat age",
            "example": 1
          },
          "breed": {
            "type": "string",
            "description": "The cat breed",
            "example": "Persian"
          },
          "color": {
            "type": "string",
            "description": "The cat color",
            "example": "White"
          },
          "ownerId": {
            "type": "number",
            "description": "The cat owner id",
            "example": 1
          }
        },
        "required": [
          "name",
          "sex",
          "age",
          "breed",
          "color",
          "ownerId"
        ]
      },
      "UpdateCatDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The cat name",
            "example": "Kitty"
          },
          "sex": {
            "type": "string",
            "description": "The cat sex",
            "example": "MALE"
          },
          "age": {
            "type": "number",
            "description": "The cat age",
            "example": 1
          },
          "breed": {
            "type": "string",
            "description": "The cat breed",
            "example": "Persian"
          },
          "color": {
            "type": "string",
            "description": "The cat color",
            "example": "White"
          },
          "ownerId": {
            "type": "number",
            "description": "The cat owner id",
            "example": 1
          }
        }
      }
    }
  }
}